feat: formulario de cliente para edicao

+++ views/cliente.php
@@ -41,7 +41,7 @@
             <td>
                 <a class="btn btn-edit" href="<?= BASE_URL ?>cliente/edit/<?= $item['id'] ?>"><i class="fas fa-pencil-alt"></i> Editar</a>
 
-                <a class="btn btn-delete" href="<?= BASE_URL ?>cliente/delete/<?= $item['id'] ?>"><i class="fas fa-minus-circle"></i> Excluir</a>
+                <a onclick="return confirm('Tem certeza que deseja excluir este cliente?')" class="btn btn-delete" href="<?= BASE_URL ?>cliente/delete/<?= $item['id'] ?>"><i class="fas fa-minus-circle"></i> Excluir</a>
             </td>
         </tr>
